Add C with libmill example.
This isn't much more than quite interesting at the moment as
libmill's current behaviour isn't suitable for this task, it
works but not amazingly, it will be more intesting to see it
libmill can and should be modified to make this work well.

I put some info in c_libmill/skynet.c to explain the situation
further.